Add mesh to ideal/default/index
app.js is calling an API called CodeMirror (codemirror.net) which
is what is giving us the text editor. CodeMirror is open source with
the MIT license which means we are free to use it.

The weird refresh behavior is also explained by app.js, there is a
function which refreshes the page once an hour.